Generate Python Code for Scanpy UMAP Density Plots

This function generates a block of Python code that, when executed, produces UMAP density plots for specified metadata variables in an AnnData object. It is designed for cross-language Seurat → Scanpy workflows, where an R object is exported to .h5ad format and processed in Python using scanpy.

density_plot(file_name, density_specs, base_figsize = c(4, 4), max_per_row = 4)

Arguments

file_name

Character string. Path to the .h5ad file that contains the AnnData object to be plotted.

density_specs

A named list specifying plotting order for each metadata variable. Each element should itself be a list containing an order vector (e.g., list(order = c("group1", "group2", ...))).

base_figsize

Numeric vector of length 2 giving the base figure size in inches for each density plot panel. Default is c(4, 4).

max_per_row

Integer specifying the maximum number of plots per row. Used to determine the number of columns in multi-panel figures.

Value

A single character string containing executable Python code. This code can be passed directly to reticulate::py_run_string().

Details

The generated Python code will:

Import required libraries (scanpy and matplotlib).

Read the .h5ad file into an AnnData object.

For each metadata variable in density_specs, compute and plot UMAP density maps using scanpy.tl.embedding_density() and scanpy.pl.embedding_density().

Save each density plot to a PNG file named density_<meta_col>.png.

See also

Examples

if (FALSE) { # \dontrun{
density_plot_specs <- list(
  cohort_tp = list(order = c("Chronic Pre-DAA", "Chronic Post-DAA")),
  protein_group = list(order = c("Group1", "Group2"))
)

py_code <- density_plot(
  file_name = "mydata.h5ad",
  density_specs = density_plot_specs
)

} # }