Generate DotPlot PDFs for a Seurat Object

This function creates a PDF containing two DotPlots for a given Seurat object: one using a set of selected features, and another using top marker genes identified by FindAllMarkers() and selected with select_marker_genes_score().

dotplots_pdf(object, features.1, features.2, file_name = "dotplot.pdf")

Arguments

object

A Seurat object.

features.1

A character vector of feature names to include in the first DotPlot.

features.2

A character vector of feature names to include in the second DotPlot.

file_name

A string giving the output pdf filename (dotplot.pdf).

Value

A list containing:

markers

The full marker table from FindAllMarkers()

genes

The selected marker genes returned by select_marker_genes_score()

Examples

if (FALSE) { # \dontrun{
dotplots_pdf(object = seurat_obj,
                     features.1 = c("GeneA", "GeneB"),
                     features.2 = c("GeneC", "GeneD"),
                     file_name = "dotplot.pdf")
} # }