Generate Clustered Heatmap for Seurat Object

Scales data for selected genes, arranges cells by cluster, and produces a heatmap with custom colors for clusters. Saves the heatmap as a PNG file.

heatmap_cell_level(
  object,
  features,
  idents,
  colors,
  split_by_cluster = T,
  cluster_rows = F,
  cluster_columns = F,
  file_name = NULL,
  zlim = c(-2, 2),
  width = 6,
  height = NULL,
  fontsize = 3,
  res = 300
)

Arguments

object

A Seurat object.

features

A character vector or a data frame with a column gene containing gene names to plot.

idents

Character. Column name in meta.data containing identity.

colors

Character vector of colors to use for clusters.

split_by_cluster

Logical. If TRUE, splits the heatmap by cluster (default = TRUE).

cluster_rows

Logical. Whether to cluster rows (genes) in the heatmap (default = FALSE).

cluster_columns

Logical. Whether to cluster columns (cells) in the heatmap (default = FALSE).

file_name

Character. File name for the saved heatmap PNG. If NULL, defaults to "<cluster>_heatmap.png".

zlim

Numeric vector of length 2 specifying limits for Z-score color scale (default = c(-2, 2)).

width

Numeric. Width of the output PNG in inches (default = 6).

height

Numeric. Height of the output PNG in inches (default = 8).

fontsize

fontsize of rownames. Default is 3.

res

Numeric. Resolution of the output PNG in DPI (default = 300).

Value

NULL. Heatmap is saved to file.